chore(@clayui/multi-select): update MultiSelect component stories
